Changed const page to name startnig with capital letter
