--- frontend/src/app/vendors/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>vendors </div>
-  )
-}
-
-export default page
